% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/define_projections.r
\name{define_projections}
\alias{define_projections}
\title{Define equal area and decimal lat/lon projections}
\usage{
define_projections()
}
\description{
Defind the projections for i) cylindrical equal area and
ii) decimal latitude and longitude.
importFrom sp CRS
}
\author{
Kelli Faye Johnson
}

