% Generated by roxygen2 (4.0.1): do not edit by hand
\name{knitr_clean}
\alias{knitr_clean}
\title{Remove unwanted files after using knitr}
\usage{
knitr_clean(workingdir = NA)
}
\arguments{
\item{workingdir}{specifiy the working directory}
}
\description{
Remove unwanted files after using knitr
}
\author{
Kelli Faye Johnson
}

