% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_projections.R
\name{define_projections}
\alias{define_projections}
\title{Define equal projections and globally assign them.}
\usage{
define_projections()
}
\description{
Defind the projections for
  \enumerate{
    \item Alaska Albers
    \item cylindrical equal area
    \item decimal latitude and longitude.
  }
See \href{http://www.spatialreference.org/}{Spatial Reference}
}
\author{
Kelli Faye Johnson
}

