% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knitr_clean.r
\name{knitr_clean}
\alias{knitr_clean}
\title{Remove unwanted files after using knitr}
\usage{
knitr_clean(workingdir = NA)
}
\arguments{
\item{workingdir}{specifiy the working directory}
}
\description{
Remove unwanted files after using knitr
}
\author{
Kelli Faye Johnson
}

