% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_file_off.R
\name{make_file_off}
\alias{make_file_off}
\title{Turn off device created from make_file}
\usage{
make_file_off()
}
\description{
A function stolen from Cole Monnahan
}
\author{
Cole C. Monnahan

Kelli Faye Johnson
}
